+++ back-end/app/Http/Controllers/MainTaskController.php
@@ -4,8 +4,10 @@ namespace App\Http\Controllers;
 
 
 use App\Models\MainTask;
+use App\Models\SubTask;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Illuminate\Http\Request;
+use Tymon\JWTAuth\Facades\JWTAuth;
 use function Laravel\Prompts\error;
 
 class MainTaskController extends Controller
@@ -13,12 +15,14 @@ class MainTaskController extends Controller
     /**
      * Display a listing of the resource.
      */
-    public function index(string $id)
+    public function index()
     {
+        $userId = JWTAuth::parseToken()->authenticate()->id;
+
         return MainTask::with(['group', 'users', 'subTasks' => function ($query) {
             $query->orderBy('completed', 'asc')->orderBy('created_at', 'asc');
-        }])->whereHas('users', function ($query) use ($id) {
-            $query->where('users.id', $id);
+        }])->whereHas('users', function ($query) use ($userId) {
+            $query->where('users.id', $userId);
         })->orderBy('deadline', 'asc')->get();
     }
 
@@ -27,11 +31,15 @@ class MainTaskController extends Controller
      */
     public function create(request $request)
     {
-        if (!$request->title || !$request->deadline || !$request->ai_file) {
-            return response(['error' => 'you are stupid'], 404);
-        }
-
         try {
+            if (!$request->title || !$request->deadline || !$request->ai_file) {
+                return response(['error' => 'you are stupid'], 404);
+            }
+            $aiFile = $request->file('ai_file')->storePublicly('storage', 'public');
+
+// todo aanpassen naar admin ipv user_id
+            $userId = JWTAuth::parseToken()->authenticate()->id;
+
             $aiFile = $request->file('ai_file')->storePublicly('storage', 'public');
 
             $mainTask = new MainTask([
@@ -40,11 +48,10 @@ class MainTaskController extends Controller
                 'description' => $request->description,
                 'ai_file' => $aiFile,
                 'group_id' => $request->group_id,
-
             ]);
             $mainTask->save();
 
-            $mainTask->users()->attach($request->user_id, [
+            $mainTask->users()->attach($userId, [
                 'level' => $request->level ?? "beginner",
                 'progress' => $request->progress ?? 0,
                 'score' => $request->score ?? null,
@@ -62,9 +69,17 @@ class MainTaskController extends Controller
      */
     public function show(string $id)
     {
-        return MainTask::with(['group', 'users', 'subTasks' => function ($query) {
-            $query->orderBy('completed', 'asc')->orderBy('created_at', 'asc');
-        }])->findOrFail($id);
+        $userId = JWTAuth::parseToken()->authenticate()->id;
+
+        $mainTask = MainTask::with('users')->findOrFail($id);
+
+        if ($mainTask->users->contains('id', $userId)) {
+            return MainTask::with(['group', 'users', 'subTasks' => function ($query) {
+                $query->orderBy('completed', 'asc')->orderBy('created_at', 'asc');
+            }])->findOrFail($id);
+        } else {
+            return response()->json(['error' => 'you are not authorized'], 403);
+        }
     }
 
     /**
@@ -77,25 +92,32 @@ class MainTaskController extends Controller
             if (!$request->title || !$request->deadline || !$request->ai_file) {
                 return response(['error' => 'you are stupid'], 404);
             }
+
             $mainTask = MainTask::query()->findOrFail($id);
+            $userId = JWTAuth::parseToken()->authenticate()->id;
+            // todo aanpassen naar admin ipv user_id
+            if ($mainTask->user_id === $userId) {
+                $mainTask->update([
+                    'title' => $request->title ?? $mainTask->title,
+                    'deadline' => $request->deadline ?? $mainTask->deadline,
+                    'description' => $request->description ?? $mainTask->description,
+                    'ai_file' => $request->ai_file ?? $mainTask->ai_file,
+                    'group_id' => $request->group_id ?? $mainTask->group_id,
 
-            // is kinda redundant, but we'll leave it there for safety
-            $mainTask->update([
-                'title' => $request->title ?? $mainTask->title,
-                'deadline' => $request->deadline ?? $mainTask->deadline,
-                'description' => $request->description ?? $mainTask->description,
-                'ai_file' => $request->ai_file ?? $mainTask->ai_file,
-                'group_id' => $request->group_id ?? $mainTask->group_id,
+                ]);
+                $mainTask->save();
 
-            ]);
-            $mainTask->save();
+                $mainTask->users()->updateExistingPivot($userId, [
+                    'level' => $request->level ?? $mainTask->level,
+                    'progress' => $request->progress ?? $mainTask->progress,
+                    'score' => $request->score ?? $mainTask->score,
+                ]);
+                return $mainTask;
+            } else {
+                return response()->json(['error' => 'you are not authorized'], 403);
+            }
 
-            $mainTask->users()->updateExistingPivot($request->user_id, [
-                'level' => $request->level ?? $mainTask->level,
-                'progress' => $request->progress ?? $mainTask->progress,
-                'score' => $request->score ?? $mainTask->score,
-            ]);
-            return $mainTask;
+
         } catch (ModelNotFoundException $e) {
             return response(['error' => $e], 500);
         }
@@ -107,8 +129,17 @@ class MainTaskController extends Controller
      */
     public function delete(string $id)
     {
+
+// todo aanpassen naar admin ipv user_id
+        $userId = JWTAuth::parseToken()->authenticate()->id;
         $mainTask = MainTask::query()->findOrFail($id);
-        $mainTask->delete();
-        return $mainTask;
+
+        if ($mainTask->user_id === $userId) {
+            $mainTask->delete();
+            return $mainTask;
+        } else {
+            return response()->json(['error' => 'you are not authorized'], 403);
+        }
     }
 }
+// ToDo: file laden op websites
